requirements.txt
tiktoken==0.8.0
llama-index==0.12.10.post1
chromadb==0.6.2
python-dotenv==1.0.1
transformers==4.47.1
torch==2.5.1
torchvision==0.20.1
sentence-transformers==3.3.1
streamlit==1.41.1
requests==2.32.3
python-docx==1.1.2
fpdf==1.7.2